gummfit-agency/server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
email-validator>=2.0.0
httpx>=0.25.0
boto3>=1.34.0
redis>=5.0.0
python-multipart>=0.0.6
